Add toDos to list
The todos are now added to the screen after clicking the add button. They are removed from the screen and the state is when the trash button is clicked.

When the complete button is clicked, the completeHandler toggles the state  of the 'completed' property of the todo to true/false. In the next commit I will work on linking this to update the ui of the todo item, showing it to be complete.